Move src/families files to reject mobile/common approach
