+++ .prettierignore
@@ -1,3 +1,4 @@
 package.json
 test-e2e/**/*.json
 static/**/*.json
+src/generated
