+++ src/components/modals/Send/SendRecipientFields.js
@@ -1,15 +1,7 @@
 // @flow
 import React from 'react'
 import type { Account, Transaction, TransactionStatus } from '@ledgerhq/live-common/lib/types'
-import BitcoinKind from './BitcoinKind'
-import EthereumKind from './EthereumKind'
-import RippleKind from './RippleKind'
-
-const byFamily = {
-  bitcoin: BitcoinKind,
-  ethereum: EthereumKind,
-  ripple: RippleKind,
-}
+import byFamily from 'generated/SendRecipientFields'
 
 type Props = {
   account: Account,
@@ -18,10 +10,10 @@ type Props = {
   onChange: Transaction => void,
 }
 
-const FeeField = (props: Props) => {
+const RecipientRelatedField = (props: Props) => {
   const Cmp = byFamily[props.account.currency.family]
   if (!Cmp) return null
   return <Cmp {...props} />
 }
 
-export default FeeField
+export default RecipientRelatedField
